+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * getETag from getObjectMetadata returns null ([#48](https://github.com/findify/s3mock/issues/48))
 * update to akka 2.5.2, akka-http 10.0.7
 * fix concurrent requests causing weird locking issues on FileProvider ([#52](https://github.com/findify/s3mock/issues/52))
+* fix warnings in GetObject about incorrent headers ([#54](https://github.com/findify/s3mock/issues/54))
 
 0.2.2
 =======